update symlink script

+++ apps/AWS/syntrillo-clinic-backend/utils/deployments/symlinks-recreate.sh
@@ -5,16 +5,22 @@ create_symlinks() {
     
     cd "$dir"
     
+    echo "symlink rewrite => $dir ($2)"
+
     # Array of files/dirs to process
-    local items=("routes" "static" "syntrillo" "templates" "api.py") 
+    local items=($2) 
 
     # Process each item
     for item in "${items[@]}"; do
         if [ -e "$item" ]; then
-            # echo "Processing $item"
-            mv "$item" "${item}.tmp"
-            ln -s "$(cat ${item}.tmp)" "$item"
-            rm "${item}.tmp"
+            if [ -n "$CODEBUILD_BUILD_ID" ]; then # only execute this in CodeBuild
+                mv "$item" "${item}.tmp"
+                ln -s "$(cat ${item}.tmp)" "$item"
+                rm "${item}.tmp"
+            else
+                echo "Processing $item"
+            fi            
+
         fi
     done
